feat(agent-supervisor): ASE3-009 production Python Supervisor facade
Add lazy Supervisor.open()/run/preview/steer/status/follow/explain/doctor
API with body-free ProductionServiceCompositionManifest bound to ASE3-026
activation. Export from entrypoints, agent_supervisor, and package root
without cold-import side effects. Relocate activation gate validators under
validation so domain packages do not import entrypoints.
