+++ ipfs_accelerate_py/__init__.py
@@ -975,11 +975,58 @@ __all__ = [
 ]
 
 
+def _load_supervisor_facade(name: str) -> Any:
+    """ASE3-009 lazy product facade (cold-safe; no process start)."""
+
+    from .agent_supervisor.entrypoints.facade import (
+        Supervisor,
+        SupervisorAmbiguityError,
+        SupervisorConfigurationError,
+        SupervisorError,
+        SupervisorObservation,
+        SupervisorRun,
+        SupervisorUnavailableError,
+    )
+
+    exports = {
+        "Supervisor": Supervisor,
+        "SupervisorRun": SupervisorRun,
+        "SupervisorObservation": SupervisorObservation,
+        "SupervisorError": SupervisorError,
+        "SupervisorConfigurationError": SupervisorConfigurationError,
+        "SupervisorAmbiguityError": SupervisorAmbiguityError,
+        "SupervisorUnavailableError": SupervisorUnavailableError,
+    }
+    if name not in exports:
+        raise AttributeError(name)
+    value = exports[name]
+    _PUBLIC_VALUES[name] = value
+    _PUBLIC_RESOLVED.add(name)
+    # Cache on the module so subsequent ModuleType lookups succeed.
+    globals()[name] = value
+    return value
+
+
+_SUPERVISOR_FACADE_EXPORTS = frozenset(
+    {
+        "Supervisor",
+        "SupervisorRun",
+        "SupervisorObservation",
+        "SupervisorError",
+        "SupervisorConfigurationError",
+        "SupervisorAmbiguityError",
+        "SupervisorUnavailableError",
+    }
+)
+
+
 def __getattr__(name: str) -> Any:
     if name == "worker":
         return _load_legacy_worker()
     if name in _NAME_TO_GROUP:
         return _resolve_public(name)
+    if name in _SUPERVISOR_FACADE_EXPORTS:
+        return _load_supervisor_facade(name)
     if name == "cli_main" and SKIP_CORE:
         return None
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
@@ -995,6 +1042,11 @@ class _IPFSAccelerateModule(ModuleType):
         namespace = ModuleType.__getattribute__(self, "__dict__")
         if name in namespace.get("_NAME_TO_GROUP", ()):
             return namespace["_resolve_public"](name)
+        if name in namespace.get("_SUPERVISOR_FACADE_EXPORTS", ()):
+            try:
+                return ModuleType.__getattribute__(self, name)
+            except AttributeError:
+                return namespace["_load_supervisor_facade"](name)
         return ModuleType.__getattribute__(self, name)
 
     def __setattr__(self, name: str, value: Any) -> None:
