+++ test/api/test_agent_supervisor_entrypoint_package.py
@@ -27,7 +27,22 @@ def test_reviewed_contract_exports_preserve_exact_object_identity() -> None:
         "ipfs_accelerate_py.agent_supervisor.entrypoints"
     )
     assert entrypoints.ENTRYPOINT_CONTRACT_EXPORTS == tuple(contracts.__all__)
-    assert entrypoints.ENTRYPOINT_LAZY_FACADE_EXPORTS == ()
+    assert entrypoints.ENTRYPOINT_LAZY_FACADE_EXPORTS == (
+        "Supervisor",
+        "SupervisorRun",
+        "SupervisorObservation",
+        "SupervisorError",
+        "SupervisorConfigurationError",
+        "SupervisorAmbiguityError",
+        "SupervisorUnavailableError",
+        "ProductionServiceCompositionManifest",
+        "ProductionServiceComposition",
+        "resolve_production_composition",
+        "build_production_composition_manifest",
+        "ServiceCompositionError",
+        "ActivationNotReadyError",
+        "ConfigurationUnavailableError",
+    )
 
     metadata = {
         "ENTRYPOINT_PACKAGE_NAME",
@@ -35,19 +50,23 @@ def test_reviewed_contract_exports_preserve_exact_object_identity() -> None:
         "ENTRYPOINT_LAZY_FACADE_EXPORTS",
         "ENTRYPOINT_LOWER_DOMAIN_PACKAGES",
     }
-    assert set(entrypoints.__all__) == metadata | set(contracts.__all__)
+    assert set(entrypoints.__all__) == (
+        metadata | set(contracts.__all__) | set(entrypoints.ENTRYPOINT_LAZY_FACADE_EXPORTS)
+    )
     for name in contracts.__all__:
         assert getattr(entrypoints, name) is getattr(contracts, name)
 
     # Reloading the composition package does not manufacture replacement
-    # classes/enums or eagerly resolve a future facade.
+    # classes/enums or eagerly resolve a facade.
     request_type = entrypoints.SupervisorInvocationRequest
     reloaded = importlib.reload(entrypoints)
     assert reloaded.SupervisorInvocationRequest is request_type
     assert not hasattr(reloaded, "SupervisorControlService")
     assert not hasattr(reloaded, "PromptSupervisorService")
     assert not hasattr(reloaded, "PortalImplementationSupervisor")
-
+    # Lazy facade symbols are listed but not eagerly bound as live class objects
+    # until first attribute access.
+    assert not isinstance(vars(reloaded).get("Supervisor"), type)
 
 def test_import_is_cold_provider_free_and_effect_free(tmp_path: Path) -> None:
     script = r"""
@@ -140,13 +159,33 @@ print(json.dumps({
     payload = json.loads(completed.stdout)
     assert payload["added_forbidden"] == []
     assert payload["before_files"] == payload["after_files"] == []
-    assert payload["lazy_facades"] == []
+    assert payload["lazy_facades"] == [
+        "Supervisor",
+        "SupervisorRun",
+        "SupervisorObservation",
+        "SupervisorError",
+        "SupervisorConfigurationError",
+        "SupervisorAmbiguityError",
+        "SupervisorUnavailableError",
+        "ProductionServiceCompositionManifest",
+        "ProductionServiceComposition",
+        "resolve_production_composition",
+        "build_production_composition_manifest",
+        "ServiceCompositionError",
+        "ActivationNotReadyError",
+        "ConfigurationUnavailableError",
+    ]
     assert payload["has_service"] is False
     assert {
         "ipfs_accelerate_py.agent_supervisor.entrypoints",
         "ipfs_accelerate_py.agent_supervisor.entrypoints.contracts",
     } <= set(payload["added"])
-
+    # Cold import must not eagerly load facade/service_factory modules.
+    assert "ipfs_accelerate_py.agent_supervisor.entrypoints.facade" not in payload["added"]
+    assert (
+        "ipfs_accelerate_py.agent_supervisor.entrypoints.service_factory"
+        not in payload["added"]
+    )
 
 def _node_imports_entrypoints(node: ast.AST) -> bool:
     target = "ipfs_accelerate_py.agent_supervisor.entrypoints"
@@ -222,6 +261,7 @@ def test_entrypoint_initializer_imports_contracts_only() -> None:
             imported.add(("." * node.level) + (node.module or ""))
 
     assert imported == {"__future__", "typing", "."}
+    # Lazy facade resolution is allowed via __getattr__, not eager imports.
 
 
 def test_package_map_documents_highest_layer_and_storage_boundary() -> None:
